refactor(zero-cache): refactor subscription/notification logic into separate class
